requirements.txt
# CCD Orphanage Portal — Django Backend
# Python 3.11+

Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-environ>=0.11,<1.0
django-storages[s3]>=1.14,<2.0
boto3>=1.34,<2.0
gunicorn>=21.0,<23.0
uvicorn[standard]>=0.30,<1.0
pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
requests>=2.31,<3.0
twilio>=8.0,<9.0
africastalking>=2.0,<3.0
django-ratelimit>=4.0,<5.0

# Dev / linting
ruff>=0.8,<1.0
black>=24.0,<25.0
